refactoring user logic to write packages
