+++ src/main/java/com/isa/morswiny/Dao/FavouritesDao.java
@@ -1,0 +1,4 @@
+package com.isa.morswiny.Dao;
+
+public class FavouritesDao {
+}
