+++ src/main/java/com/isa/morswiny/Dao/TicketDao.java
@@ -1,4 +1,4 @@
-package com.isa.morswiny.eventsDao;
+package com.isa.morswiny.Dao;
