+++ src/main/java/com/isa/morswiny/dto/FavouritesDto.java
@@ -1,0 +1,4 @@
+package com.isa.morswiny.dto;
+
+public class FavouritesDto {
+}
