+++ src/main/java/com/isa/morswiny/repository/EventRepository.java
@@ -1,6 +1,6 @@
 package com.isa.morswiny.repository;
 
-import com.isa.morswiny.eventsDao.EventDao;
+import com.isa.morswiny.Dao.EventDao;
 import com.isa.morswiny.model.Event;
 
 import javax.enterprise.context.RequestScoped;
