+++ src/main/java/com/isa/morswiny/services/EventDbLoadService.java
@@ -1,6 +1,6 @@
 package com.isa.morswiny.services;
 
-import com.isa.morswiny.eventsDao.EventDao;
+import com.isa.morswiny.Dao.EventDao;
 import com.isa.morswiny.model.Event;
 
 import javax.enterprise.context.RequestScoped;
