+++ src/main/java/com/isa/morswiny/services/EventService.java
@@ -1,7 +1,7 @@
 package com.isa.morswiny.services;
 
 import com.isa.morswiny.dto.EventDto;
-import com.isa.morswiny.eventsDao.EventSearchRepositoryInterface;
+import com.isa.morswiny.Dao.EventSearchRepositoryInterface;
 import com.isa.morswiny.model.Event;
 
 import javax.inject.Inject;
