+++ src/main/java/com/isa/morswiny/servlets/DeleteEventServlet.java
@@ -1,6 +1,6 @@
 package com.isa.morswiny.servlets;
 
-import com.isa.morswiny.eventsDao.EventCRUDRepositoryInterface;
+import com.isa.morswiny.Dao.EventCRUDRepositoryInterface;
 import com.isa.morswiny.freemarker.TemplateProvider;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
