+++ src/main/java/com/isa/morswiny/servlets/SearchEventsServlet.java
@@ -1,9 +1,9 @@
 package com.isa.morswiny.servlets;
 
-import com.isa.morswiny.eventsDao.EventDao;
+import com.isa.morswiny.Dao.EventDao;
 import com.isa.morswiny.model.Event;
-import com.isa.morswiny.eventsDao.EventCRUDRepositoryInterface;
-import com.isa.morswiny.eventsDao.EventSearchRepositoryInterface;
+import com.isa.morswiny.Dao.EventCRUDRepositoryInterface;
+import com.isa.morswiny.Dao.EventSearchRepositoryInterface;
 import com.isa.morswiny.freemarker.TemplateProvider;
 import com.isa.morswiny.repository.EventRepository;
 import freemarker.template.Template;
