+++ src/main/java/com/isa/morswiny/servlets/SearchEventsServletBeta.java
@@ -1,8 +1,8 @@
 package com.isa.morswiny.servlets;
 
 import com.isa.morswiny.model.Event;
-import com.isa.morswiny.eventsDao.EventCRUDRepositoryInterface;
-import com.isa.morswiny.eventsDao.EventSearchRepositoryInterface;
+import com.isa.morswiny.Dao.EventCRUDRepositoryInterface;
+import com.isa.morswiny.Dao.EventSearchRepositoryInterface;
 import com.isa.morswiny.freemarker.TemplateProvider;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
