+++ src/test/java/com/isa/morswiny/Dao/EventSearchRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.isa.morswiny.eventsDao;
+package com.isa.morswiny.Dao;
